+++ js_libraries/types/test/pages/index.tsx
@@ -78,3 +78,35 @@ function invoke<T extends keyof UIMethods>(_param: UIMethods[T]) {}
     },
   });
 }
+
+// fetch api
+{
+  const streamToArrayBuffer = async (stream: ReadableStream) => {
+    const reader = stream.getReader();
+    while (true) {
+      const { done, value } = await reader.read();
+      if (done) {
+        break;
+      } else {
+        const text = TextCodecHelper.decode(value);
+      }
+    }
+  };
+
+  lynx
+    .fetch('url', {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        a: 1,
+      }),
+      lynxExtension: {
+        useStreaming: true,
+      },
+    })
+    .then((response) => {
+      streamToArrayBuffer(response.body);
+    });
+}