+++ js_libraries/types/types/background-thread/index.d.ts
@@ -9,3 +9,4 @@ export * from './nodes-ref';
 export * from './performance';
 export * from './native-modules';
 export * from './lynx-performance-entry';
+export * from './text-encoder-decoder';