+++ js_libraries/types/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 3.4.5
+- fix `TextCodecHelper` defines.
+- Update `fetch` defines for chunk streaming.
+
 ## 3.4.4
 - Complete the documentation for text related APIs.